git-svn-id: https://svn.tdr.iem.uni-due.de:8443/svn/sctp/trunk/rsplib2@802 516812f1-d3fd-0310-9ea4-9aededebcb87

+++ rsplib/sessionstorage.c
@@ -51,7 +51,7 @@ static void sessionAssocIDPrint(const void* node, FILE* fd)
 {
    const struct Session* session = (const struct Session*)getSessionFromAssocIDStorageNode((void*)node);
 
-   fprintf(fd, "A%u[S%u] ", session->AssocID, session->SessionID);
+   fprintf(fd, "A%u[S%u] ", (unsigned int)session->AssocID, session->SessionID);
 }
